+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerMapping/HandleInterceptor.java
@@ -1,0 +1,19 @@
+package cn.haitaoss.tinyspringmvc.framework.handlerMapping;
+
+import org.aopalliance.intercept.MethodInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/**
+ * @author haitao.chen
+ * email haitaoss@aliyun.com
+ * date 2021-04-24 17:05
+ */
+public interface HandleInterceptor extends MethodInterceptor {
+    boolean preHandle(HttpServletRequest var1, HttpServletResponse var2, Object var3) throws Exception;
+
+    // void postHandle(HttpServletRequest var1, HttpServletResponse var2, Object var3, ModelAndView var4) throws Exception;
+
+    // void afterCompletion(HttpServletRequest var1, HttpServletResponse var2, Object var3, Exception var4) throws Exception;
+}