+++ src/main/java/cn/haitaoss/tinyspringmvc/webappProject/controller/PersonController.java
@@ -1,0 +1,25 @@
+package cn.haitaoss.tinyspringmvc.webappProject.controller;
+
+import cn.haitaoss.tinyioc.beans.annotation.Autowired;
+import cn.haitaoss.tinyioc.beans.annotation.Controller;
+import cn.haitaoss.tinyspringmvc.webappProject.service.PersonService;
+
+/**
+ * @author haitao.chen
+ * email haitaoss@aliyun.com
+ * date 2021-04-24 17:20
+ *
+ */
+@Controller
+public class PersonController {
+    @Autowired
+    private PersonService personService;
+
+    public PersonService getPersonService() {
+        return personService;
+    }
+
+    public void setPersonService(PersonService personService) {
+        this.personService = personService;
+    }
+}